include: sbi: Make struct sbi_scratch to support Zcheripurecap ABI
- Use `uintptr_t` for storing pointer to make the data alignment
  compatible with both of purecap pointer or integer pointer mode.
- Reorganize `struct sbi_scratch` to make sure all member data are
  aligned to CHERI capability in Zcheripurecap ABI and integer
  in non-CHERI. The `struct sbi_scratch` is also aligned to 16 bytes
  for stack requirment.
- Add per hart stack for Zcheripurecap ABI as the stack pointer is
  no longer to per hart scratch pointer in Zcheripurecap ABI.
- Sync up the type where the code access `struct sbi_scratch`

+++ include/sbi/sbi_trap.h
@@ -261,7 +261,7 @@ static inline struct sbi_trap_context *sbi_trap_get_context(struct sbi_scratch *
 static inline void sbi_trap_set_context(struct sbi_scratch *scratch,
 					struct sbi_trap_context *tcntx)
 {
-	scratch->trap_context = (unsigned long)tcntx;
+	scratch->trap_context = (uintptr_t)tcntx;
 }
 
 struct sbi_trap_context *sbi_trap_handler(struct sbi_trap_context *tcntx);
